+++ app.js
@@ -2,8 +2,8 @@ import express from 'express';
 import session from 'express-session';
 import cors from 'cors';
 import dotenv from 'dotenv';
-import { userRoutes, notFoundRoute } from './routes';
-import errorHandler from './middlewares';
+import { userRoutes } from './routes';
+import { errorHandler, notFoundRoute } from './middlewares';
 
 dotenv.config();
 
@@ -28,7 +28,7 @@ app.use(
 );
 
 app.use('/api/v1/user', userRoutes);
-app.use('*', notFoundRoute);
+app.use(notFoundRoute);
 app.use(errorHandler);
 
 export default app;
