+++ routes/index.js
@@ -1,2 +1,2 @@
+// eslint-disable-next-line import/prefer-default-export
 export { default as userRoutes } from './userRoutes';
-export { default as notFoundRoute } from './notFoundRoute';
